# Reduce memory usage of ActiveSeriesTracker