--- a/pkg/ingester/activeseries/active_series.go
+++ b/pkg/ingester/activeseries/active_series.go
@@ -17,7 +17,7 @@ import (
 )
 
 const (
-	numStripes = 512
+	numStripes = 128
 
 	EnabledFlag     = "ingester.active-series-metrics-enabled"
 	IdleTimeoutFlag = "ingester.active-series-metrics-idle-timeout"