#### What this PR does

Three changes:
* Store stripe counts as uint32 - a range of 4 billion ought to be enough, on one ingester.
* Smaller slice for match counts - we prefer to optimise for all the times when matchers are not increasing. On a multi-tenant system this will be the vast majority.
* Reduce stripes from 512 to 128 - this reduces memory for small tenants by 75%. Benchmarks show no difference in access speed.

And extend the benchmark to include matchers.

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester/activeseries
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
                                                                 │   before.txt   │              after.txt              │
                                                                 │     sec/op     │    sec/op     vs base               │
ActiveSeriesTest_single_series/50-4                                  55.46n ±  2%   54.00n ±  6%        ~ (p=0.093 n=6)
ActiveSeriesTest_single_series/100-4                                 54.98n ±  1%   53.79n ± 10%        ~ (p=0.065 n=6)
ActiveSeriesTest_single_series/500-4                                 55.02n ±  3%   53.52n ±  6%        ~ (p=0.093 n=6)
ActiveSeries_UpdateSeries/rounds=0_series=0_matchers=0-4            112.57µ ± 38%   19.84µ ±  6%  -82.38% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=0_series=0_matchers=100-4         1052.40µ ± 12%   97.60µ ±  8%  -90.73% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=100000_matchers=0-4        42.45m ± 10%   41.85m ±  7%        ~ (p=0.818 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=100000_matchers=100-4      569.2m ± 12%   538.5m ±  2%        ~ (p=0.065 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=1000000_matchers=0-4       651.5m ± 10%   635.1m ±  5%        ~ (p=0.394 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=100000_matchers=0-4       144.4m ± 11%   137.0m ± 32%        ~ (p=0.485 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=1000000_matchers=0-4       2.125 ± 18%    2.032 ±  5%        ~ (p=0.132 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=100000_matchers=100-4     669.1m ± 19%   635.1m ±  2%   -5.08% (p=0.026 n=6)
ActiveSeries_Active_once-4                                           262.0µ ±  3%   217.9µ ±  4%  -16.84% (p=0.002 n=6)
ActiveSeries_Active_twice-4                                          280.2µ ±  8%   225.2µ ±  3%  -19.64% (p=0.002 n=6)

                                                                 │    before.txt    │               after.txt               │
                                                                 │       B/op       │     B/op      vs base                 │
ActiveSeriesTest_single_series/50-4                                   0.000 ±  0%       0.000 ± 0%        ~ (p=1.000 n=6) ¹
ActiveSeriesTest_single_series/100-4                                  0.000 ±  0%       0.000 ± 0%        ~ (p=1.000 n=6) ¹
ActiveSeriesTest_single_series/500-4                                  0.000 ±  0%       0.000 ± 0%        ~ (p=1.000 n=6) ¹
ActiveSeries_UpdateSeries/rounds=0_series=0_matchers=0-4           104.00Ki ±  0%     24.62Ki ± 0%  -76.32% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=0_series=0_matchers=100-4         2792.0Ki ±  0%     180.6Ki ± 0%  -93.53% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=100000_matchers=0-4       18.25Mi ±  0%     18.10Mi ± 0%   -0.81% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=100000_matchers=100-4     20.88Mi ±  0%     18.26Mi ± 0%  -12.53% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=1000000_matchers=0-4      272.7Mi ±  0%     272.5Mi ± 0%   -0.07% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=100000_matchers=0-4      18.25Mi ±  0%     18.10Mi ± 0%   -0.80% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=1000000_matchers=0-4     272.7Mi ±  0%     272.5Mi ± 0%   -0.07% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=100000_matchers=100-4    20.86Mi ±  0%     18.26Mi ± 0%  -12.48% (p=0.002 n=6)
ActiveSeries_Active_once-4                                           123.50 ± 10%       90.00 ± 8%  -27.13% (p=0.002 n=6)
ActiveSeries_Active_twice-4                                          128.50 ±  9%       92.00 ± 8%  -28.40% (p=0.002 n=6)

                                                                 │   before.txt   │               after.txt               │
                                                                 │   allocs/op    │  allocs/op    vs base                 │
ActiveSeriesTest_single_series/50-4                                 0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=6) ¹
ActiveSeriesTest_single_series/100-4                                0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=6) ¹
ActiveSeriesTest_single_series/500-4                                0.000 ±  0%      0.000 ±  0%        ~ (p=1.000 n=6) ¹
ActiveSeries_UpdateSeries/rounds=0_series=0_matchers=0-4            513.0 ±  0%      129.0 ±  0%  -74.85% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=0_series=0_matchers=100-4         2049.0 ±  0%      513.0 ±  0%  -74.96% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=100000_matchers=0-4      107.7k ±  0%     104.3k ±  0%   -3.14% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=100000_matchers=100-4    109.2k ±  0%     104.7k ±  0%   -4.15% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=1_series=1000000_matchers=0-4     1.035M ±  0%     1.037M ±  0%   +0.15% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=100000_matchers=0-4     107.7k ±  0%     104.3k ±  0%   -3.14% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=1000000_matchers=0-4    1.035M ±  0%     1.037M ±  0%   +0.15% (p=0.002 n=6)
ActiveSeries_UpdateSeries/rounds=10_series=100000_matchers=100-4   109.2k ±  0%     104.7k ±  0%   -4.12% (p=0.002 n=6)
ActiveSeries_Active_once-4                                          5.000 ± 20%      4.000 ± 25%  -20.00% (p=0.011 n=6)
ActiveSeries_Active_twice-4                                         6.000 ± 17%      5.000 ± 20%  -16.67% (p=0.013 n=6)
```

#### Checklist

- [x] Tests updated
- NA Documentation added
- [x] `CHANGELOG.md` updated