@pracucci 2023-08-03
> > Reduce stripes from 512 to 128 - this reduces memory for small tenants by 75%. Benchmarks show no difference in access speed.
> 
> We don't even have any concurrency test exercising the stripes. I'm going to add it in a follow up PR to not block this one.

See: https://github.com/grafana/mimir/pull/5668